+++ include/CeLux/backends/cpu/Decoder.hpp
@@ -13,6 +13,7 @@ class Decoder : public celux::Decoder
         : celux::Decoder( numThreads, filters)
     {
         initialize(filePath);
+        initializeAudio();
     }
 
     // No need to override methods unless specific behavior is needed