+++ src/CeLux/python/VideoReader.cpp
@@ -35,6 +35,8 @@ VideoReader::VideoReader(const std::string& filePath, int numThreads,
             celux::Factory::createDecoder(torchDevice, filePath, numThreads, filters_);
         CELUX_INFO("Decoder created successfully");
 
+        audio = std::make_shared<Audio>(decoder); // Create audio object
+
         torch::Dtype torchDataType;
 
         torchDataType = findTypeFromBitDepth();
@@ -74,6 +76,38 @@ VideoReader::VideoReader(const std::string& filePath, int numThreads,
     }
 }
 
+std::shared_ptr<VideoReader::Audio> VideoReader::getAudio()
+{
+    return audio;
+}
+
+// -------------------------
+// Audio Class Implementation
+// -------------------------
+
+VideoReader::Audio::Audio(std::shared_ptr<celux::Decoder> decoder)
+    : decoder(std::move(decoder))
+{
+    if (!this->decoder)
+    {
+        throw std::runtime_error("Audio: Invalid decoder instance provided.");
+    }
+}
+
+torch::Tensor VideoReader::Audio::getAudioTensor()
+{
+    return decoder->getAudioTensor();
+}
+
+bool VideoReader::Audio::extractToFile(const std::string& outputFilePath)
+{
+    return decoder->extractAudioToFile(outputFilePath);
+}
+
+celux::Decoder::VideoProperties VideoReader::Audio::getProperties() const
+{
+    return decoder->getVideoProperties();
+}
 
 VideoReader::~VideoReader()
 {